backend/pipeline/requirements.txt
httpx==0.28.1
apscheduler==3.10.4
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
